Major refactoring....
Internally we now treat everything as 'shorts' to keep 0 ~ 255 values intact.

This addresses issues we've been having with java's lack of unsigned byte type. Basically, everything coming in from the socket 
is cast to a short with the value of (byte & 0xff). This is faster / better than if (byte < 0) byte +=256;

Yick.

Either way, it's a bit easier to code around now. As we can read shorts, return shorts, and let the TelnetClient handle 
transformation to / from byte[].